backend/requirements.txt
fastapi
uvicorn[standard]
python-multipart
python-dotenv
requests
pydantic
openai
google-api-python-client
google-auth
google-auth-oauthlib
google-auth-httplib2
